refactor(init): isolate pruning trie store into dedicated PruningTrieStoreModule
Extract all pruning-specific DI registrations out of WorldStateModule into a
new PruningTrieStoreModule, leaving WorldStateModule as an abstract contract
with a fail-fast stub for IWorldStateManager.

PruningTrieStoreModule and FlatWorldStateModule are now peer backends selected
via if/else in NethermindModule, eliminating the NotSupportedException override
stubs that FlatWorldStateModule previously needed.

IStateReader is now mapped from IWorldStateManager.GlobalStateReader in the
abstract module so both backends share the same mapping. IReadOnlyTrieStore is
removed from DI; callers use IWorldStateManager.CreateReadOnlyTrieStore().
