+++ src/Nethermind/Nethermind.Db.Test/StandardDbInitializerTests.cs
@@ -87,7 +87,7 @@ public class StandardDbInitializerTests
             {
                 DownloadReceiptsInFastSync = useReceipts
             }))
-            .AddModule(new WorldStateModule(initConfig)) // For the full pruning db
+            .AddModule(new PruningTrieStoreModule(initConfig)) // For the full pruning db
             .AddSingleton<IPruningConfig>(new PruningConfig())
             .AddSingleton<IDbConfig>(new DbConfig())
             .AddSingleton<IInitConfig>(initConfig)
