+++ src/Nethermind/Nethermind.Init/Modules/NethermindModule.cs
@@ -51,7 +51,7 @@ public class NethermindModule(ChainSpec chainSpec, IConfigProvider configProvide
                 configProvider.GetConfig<ISyncConfig>()
             ))
             .AddModule(new DbMonitoringModule())
-            .AddModule(new WorldStateModule(configProvider.GetConfig<IInitConfig>()))
+            .AddModule(new WorldStateModule())
             .AddModule(new PrewarmerModule(configProvider.GetConfig<IBlocksConfig>()))
             .AddModule(new BuiltInStepsModule())
             .AddModule(new DatabaseMigrationsModule())
@@ -97,6 +97,8 @@ public class NethermindModule(ChainSpec chainSpec, IConfigProvider configProvide
 
         if (configProvider.GetConfig<IFlatDbConfig>().Enabled)
             builder.AddModule(new FlatWorldStateModule(configProvider.GetConfig<IFlatDbConfig>()));
+        else
+            builder.AddModule(new PruningTrieStoreModule(configProvider.GetConfig<IInitConfig>()));
     }
 
     // Just a wrapper to make it clear, these three are expected to be available at the time of configurations.
